k pet 이전

+++ HODOO/src/main/java/net/octacomm/sample/domain/Statistics.java
@@ -17,5 +17,5 @@ public class Statistics implements Domain {
     
     private String theYear;
 
-    private int average;
+    private float average;
 }
